reorganize test
