+++ .settings/keycloak/Dockerfile
@@ -3,9 +3,7 @@ FROM ubuntu:18.04
 ARG KEYCLOAK_URL
 ARG DOCKERIZE_VERSION=v0.6.1
 
-
 ENV KEYCLOAK_URL=$KEYCLOAK_URL
-
 
 RUN echo "Acquire::Check-Valid-Until \"false\";\nAcquire::Check-Date \"false\";" | cat > /etc/apt/apt.conf.d/10no--check-valid-until
 
@@ -16,5 +14,6 @@ RUN wget https://github.com/jwilder/dockerize/releases/download/$DOCKERIZE_VERSI
     && rm dockerize-alpine-linux-amd64-$DOCKERIZE_VERSION.tar.gz
 
 COPY createuser.sh /tmp/createuser.sh
+RUN chmod +x /tmp/createuser.sh
 
 CMD ["/tmp/createuser.sh"]
