+++ tests/src/test/java/ca/bc/gov/open/jag/efiling/TestConfig.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.open;
+package ca.bc.gov.open.jag.efiling;
 
 import org.springframework.context.annotation.Configuration;
